RG-2288 Extract scroll behaviour hook
