+++ src/date-picker/months.tsx
@@ -1,4 +1,3 @@
-import {useCallback, useLayoutEffect, useRef, useState} from 'react';
 import {addMonths} from 'date-fns/addMonths';
 import {getDay} from 'date-fns/getDay';
 import {getDaysInMonth} from 'date-fns/getDaysInMonth';
@@ -7,8 +6,7 @@ import {startOfMonth} from 'date-fns/startOfMonth';
 import Month from './month';
 import MonthNames from './month-names';
 import units, {type MonthsProps, WEEK, weekdays} from './consts';
-import scheduleRAF from '../global/schedule-raf';
-import {ScrollHelper} from './scroll-helper';
+import {ScrollHelper, useScrollBehavior} from './scroll-helper';
 
 import styles from './date-picker.css';
 
@@ -24,8 +22,6 @@ const EMPTY_MONTHSBACK = 3;
 const NONEMPTY_MONTHSBACK = 2;
 const MONTHSBACK = EMPTY_MONTHSBACK + NONEMPTY_MONTHSBACK;
 
-const SCROLL_HANDLE_RESUME_DELAY = 10;
-
 function monthHeight(date: Date | number) {
   const monthStart = startOfMonth(date);
   const daysSinceLastFriday = (getDay(monthStart) + FridayToSunday) % WEEK;
@@ -40,48 +36,16 @@ const scrollHelper = new ScrollHelper({
   getItemHeight: monthHeight,
 });
 
-const scheduleScroll = scheduleRAF();
-
 export default function Months(props: MonthsProps) {
   const {scrollDate, onScroll} = props;
 
-  const [scrollerState, setScrollerState] = useState(() => scrollHelper.getState(scrollDate));
-
-  const componentRef = useRef<HTMLDivElement>(null);
-  const pauseScrollHandlingRef = useRef(false);
-
-  useLayoutEffect(() => {
-    if (!componentRef.current) return undefined;
-
-    componentRef.current.scrollTop = scrollerState.scrollTop;
-    const timeoutId = setTimeout(() => {
-      pauseScrollHandlingRef.current = false;
-    }, SCROLL_HANDLE_RESUME_DELAY);
-    return () => clearTimeout(timeoutId);
-  }, [scrollerState]);
-
-  const handleScroll = useCallback(() => {
-    scheduleScroll(() => {
-      if (pauseScrollHandlingRef.current) return;
-
-      const scrollTop = componentRef.current?.scrollTop;
-      if (scrollTop == null) return;
-
-      const newScrollDate = scrollHelper.getScrollDate(scrollerState.items, scrollTop);
-      onScroll(Number(newScrollDate));
-
-      if (!scrollHelper.isMidItem(scrollerState.items, newScrollDate)) {
-        setScrollerState(scrollHelper.getState(newScrollDate));
-        pauseScrollHandlingRef.current = true;
-      }
-    });
-  }, [onScroll, scrollerState]);
+  const {componentRef, handleScroll, items} = useScrollBehavior({scrollDate, onScroll, scrollHelper});
 
   return (
     <div className={styles.months} ref={componentRef} onScroll={handleScroll}>
       <div>
-        {scrollerState.items.map((month, i) =>
-          i < EMPTY_MONTHSBACK || i >= scrollerState.items.length - EMPTY_MONTHSBACK ? (
+        {items.map((month, i) =>
+          i < EMPTY_MONTHSBACK || i >= items.length - EMPTY_MONTHSBACK ? (
             <div style={{height: monthHeight(month)}} key={+month} />
           ) : (
             <Month {...props} month={month} key={+month} />
